+++ src/main/java/nextstep/sessions/domain/data/vo/Duration.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.data.vo;
 
 import java.time.LocalDateTime;
