+++ src/test/java/nextstep/sessions/domain/data/type/ImageTypeTest.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.data.type;
 
 import nextstep.sessions.domain.exception.SessionsException;
 import org.junit.jupiter.api.Test;
