+++ src/test/java/nextstep/sessions/domain/data/vo/ImageSizeTest.java
@@ -1,4 +1,4 @@
-package nextstep.sessions.domain;
+package nextstep.sessions.domain.data.vo;
 
 import nextstep.sessions.domain.exception.SessionsException;
 import org.junit.jupiter.api.Test;
